Sistema de armazenamento casual feito

+++ database/migrations/2025_12_28_235517_create_despensas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nome', 128);
             $table->string('marca', 128);
-            $table->decimal('quantidade');
+            $table->integer('quantidade');
             $table->string("local");
             $table->date('validade')->nullable();
         });
